% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{calculate_investment_value}
\alias{calculate_investment_value}
\title{Calculate Investment Value}
\usage{
calculate_investment_value(data, initial_investment)
}
\arguments{
\item{data}{A data frame containing Tesla stock data with columns "Date" and "Adj_Close."}

\item{initial_investment}{Numeric, the initial amount invested in Tesla stock.}
}
\value{
A numeric vector representing the investment value over time.
}
\description{
This function calculates the value of an investment in Tesla stock.
}
