% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{filter_by_date_range}
\alias{filter_by_date_range}
\title{Filter Data by Date Range}
\usage{
filter_by_date_range(data, start_date, end_date)
}
\arguments{
\item{data}{A data frame with a "Date" column.}

\item{start_date}{Date object for start of range.}

\item{end_date}{Date object for end of range.}
}
\value{
Filtered data frame.
}
\description{
Filters a data frame within a specific date range.
}
