% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{filtered_roadster_value}
\alias{filtered_roadster_value}
\title{Filtered Tesla Roadster Depreciation Data}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{date}{The year of depreciation in \code{Date} format.}
  \item{investment_value}{The depreciation value of the Tesla Roadster in USD.}
}
}
\usage{
filtered_roadster_value
}
\description{
This dataset contains yearly depreciation values for the Tesla Roadster
from 2010 to 2024. The data represents the depreciation of the vehicle
over time.
}
\examples{
data(filtered_roadster_value)
head(filtered_roadster_value)

}
\keyword{datasets}
