% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_tesla_stock}
\alias{filtered_tesla_stock}
\title{Filtered Tesla Stock Data}
\format{
A data frame with the following columns:
\describe{
  \item{Date}{The date of the stock data.}
  \item{Adj_Close}{The adjusted closing price of Tesla stock.}
  \item{investment_value}{Calculated investment value based on the adjusted close.}
}
}
\source{
Simulated for educational purposes
}
\usage{
filtered_tesla_stock
}
\description{
This dataset contains filtered data on Tesla stock prices, including date and adjusted closing prices.
}
\keyword{datasets}
