% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Launch the TeslaInvestR Shiny App}
\usage{
launch_shiny_app()
}
\value{
No return value. Launches the Shiny app.
}
\description{
Starts the Shiny app for exploring Tesla data.
}
