% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{smooth_investment_data}
\alias{smooth_investment_data}
\title{Smooth Investment Data}
\usage{
smooth_investment_data(data, value_column = "investment_value", window = 5)
}
\arguments{
\item{data}{Data frame with numeric column to smooth.}

\item{value_column}{Column name to smooth.}

\item{window}{Size of moving average window.}
}
\value{
Data frame with smoothed values.
}
\description{
Smooths fluctuations using a moving average.
}
