% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeslaInvestR-package.R
\docType{package}
\name{TeslaInvestR-package}
\alias{TeslaInvestR}
\alias{TeslaInvestR-package}
\title{TeslaInvestR: Tesla Stock vs. Car Investment Analysis}
\description{
The TeslaInvestR package provides historical data and tools for analyzing Tesla stock trends and Tesla Roadster depreciation. The package includes datasets on Tesla stock prices and Roadster value depreciation, and provides functions for easy comparison.
}
\author{
\strong{Maintainer}: Nahar Namdhari \email{nnam0004@student.monash.edu}

}
\keyword{internal}
