% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_output_function.R
\name{TeslaViz}
\alias{TeslaViz}
\title{Launch the Tesla Investment Analysis Shiny App}
\usage{
TeslaViz()
}
\value{
No return value. This function launches the Shiny app.
}
\description{
This function opens a Shiny application to compare Tesla stock prices and Roadster depreciation.
}
