% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filtered_roadster_value}
\alias{filtered_roadster_value}
\title{Filtered Tesla Roadster Depreciation Data}
\format{
A data frame with the following columns:
\describe{
  \item{date}{The date of the depreciation value.}
  \item{investment_value}{The depreciation value of the Tesla Roadster.}
}
}
\source{
Simulated for educational purposes
}
\usage{
filtered_roadster_value
}
\description{
This dataset represents the depreciation of Tesla Roadster values over time.
}
\keyword{datasets}
