% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\docType{data}
\name{filtered_tesla_stock}
\alias{filtered_tesla_stock}
\title{Filtered Tesla Stock Data}
\format{
A data frame with 3567 rows and 8 variables:
\describe{
  \item{Date}{The date of the stock price in \code{Date} format.}
  \item{Adj_Close}{The adjusted closing price of Tesla stock.}
  \item{investment_value}{The investment value calculated based on an initial investment.}
}
}
\source{
Yahoo Finance
}
\usage{
filtered_tesla_stock
}
\description{
This dataset contains filtered historical Tesla stock data, which includes
adjusted closing prices and calculated investment value based on an initial
investment of 28,385 shares. The dataset is filtered from 2011 to 2023.
}
\examples{
data(filtered_tesla_stock)
head(filtered_tesla_stock)

}
\keyword{datasets}
